requirements.txt
requests>=2.31
typer>=0.12

# Testing and document validation
pytest>=8.0.0
python-docx>=1.1.0
lxml>=5.0.0